+++ src/candlewick/multibody/RobotScene.h
@@ -72,8 +72,13 @@ namespace multibody {
     };
     static constexpr size_t kNumPipelineTypes =
         magic_enum::enum_count<PipelineType>();
-    enum VertexUniformSlots : Uint32 { TRANSFORM = 0 };
-    enum FragmentUniformSlots : Uint32 { MATERIAL, LIGHTING, SSAO_FLAG };
+    enum VertexUniformSlots : Uint32 { TRANSFORM, LIGHT_MATRICES };
+    enum FragmentUniformSlots : Uint32 {
+      MATERIAL,
+      LIGHTING,
+      SSAO_FLAG,
+      ATLAS_INFO
+    };
     enum FragmentSamplerSlots { SHADOW_MAP_SLOT, SSAO_SLOT };
 
     /// Map hpp-fcl/coal collision geometry to desired pipeline type.